% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frankenply.R
\name{frankenply}
\alias{frankenply}
\alias{?}
\title{Documentation Shortcuts, or frankenply loops}
\usage{
`?`(e1, e2)
}
\arguments{
\item{e1}{lhs}

\item{e2}{rhs}
}
\description{
see `help("Question", "utils")` for former usage. To use frankenply loops,
  just prefix  the arguments you want to loop on with `?~` if you want a list output,
  or `?~~` if you want a simplified output. You can loop on several argument but
  with only one prefix.
}
